disturbed_config/requirements.txt
django
djangorestframework==3.14.0
python-dotenv==0.21.0
gunicorn
django-reversion
django-filter
